initial fix for cli
