+++ README.md
@@ -125,8 +125,10 @@ See the [Observability documentation](https://railtownai.github.io/railtracks/ob
 <summary><b>Installation</b></summary>
 
 ```bash
-pip install railtracks 'railtracks[cli]'
+pip install railtracks 'railtracks[visual]'
 ```
+
+(`railtracks[cli]` is an alias for the same optional dependencies, for backward compatibility.)
 </details>
