+++ packages/railtracks/src/railtracks/cli/constants.py
@@ -1,0 +1,11 @@
+# Shared CLI constants (no FastAPI / heavy imports).
+
+cli_name = "railtracks"
+cli_directory = ".railtracks"
+UI_VERSION_FILE = f"{cli_directory}/.ui_version"
+DEFAULT_PORT = 3030
+
+# TODO: Once we are releasing to PyPi change this to the release asset instead
+latest_ui_url = (
+    "https://railtownazureb2c.blob.core.windows.net/cdn/rc-viz/latest.zip"
+)
