+++ packages/railtracks/src/railtracks/human_in_the_loop/__init__.py
@@ -1,4 +1,22 @@
+from __future__ import annotations
+
 from .human_in_the_loop import HIL, HILMessage
-from .local_chat_ui import ChatUI
 
 __all__ = ["HIL", "HILMessage", "ChatUI"]
+
+
+def __getattr__(name: str):
+    if name == "ChatUI":
+        try:
+            from .local_chat_ui import ChatUI
+
+            return ChatUI
+        except ImportError as e:
+            from railtracks.visual_extra import VisualExtraRequired
+
+            raise VisualExtraRequired(
+                "The local chat UI requires optional dependencies. "
+                "Install with: pip install 'railtracks[visual]' "
+                "(or 'railtracks[cli]' for backward compatibility)."
+            ) from e
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
