+++ packages/railtracks/src/railtracks/interaction/__init__.py
@@ -1,7 +1,8 @@
+from __future__ import annotations
+
 from ._call import call
 from .batch import call_batch
 from .broadcast_ import broadcast
-from .interactive import local_chat
 
 __all__ = [
     "call",
@@ -9,3 +10,15 @@ __all__ = [
     "broadcast",
     "local_chat",
 ]
+
+
+def __getattr__(name: str):
+    if name == "local_chat":
+        from .interactive import local_chat
+
+        return local_chat
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+
+
+def __dir__() -> list[str]:
+    return sorted(set(__all__))
